Diagnostic printing in FieldPropagationHandler
Fixes in magneticfield/tests

Disabled non-compiling tests. Minor fixes for gcc.

Compilation fixes to adapt to changed in GeantTrack. Disabled usage of non-compiling CMSFieldConstruction.

+++ magneticfield/inc/SimpleIntegrationDriver.h
@@ -625,7 +625,7 @@ SimpleIntegrationDriver<T_Stepper, Nvar>
    fSmallestFraction( right.fSmallestFraction ),
    // fNoIntegrationVariables( right.fNoIntegrationVariables ),
    fMinNoVars( right.fMinNoVars ),
-   fNoVars( std::max( Nvar, fMinNoVars )),
+   fNoVars( std::max( (int)Nvar, fMinNoVars )),
    fPowerShrink( right.fPowerShrink ),
    fPowerGrow(   right.fPowerGrow),
    fErrcon( right.fErrcon ),
